enhance address unit test
