+++ src/test/scala/vsys/account/AddressOrAliasSpecification.scala
@@ -4,7 +4,7 @@ import org.scalatest.prop.{GeneratorDrivenPropertyChecks, PropertyChecks}
 import org.scalatest.{Matchers, PropSpec}
 import vsys.blockchain.transaction.TransactionGen
 
-class AccountOrAliasSpecification extends PropSpec with PropertyChecks with GeneratorDrivenPropertyChecks with Matchers with TransactionGen {
+class AddressOrAliasSpecification extends PropSpec with PropertyChecks with GeneratorDrivenPropertyChecks with Matchers with TransactionGen {
 
   property("Account serialization round trip") {
     forAll(accountGen) { account: PrivateKeyAccount =>
