+++ src/test/scala/vsys/account/AddressOrAliasTests.scala
@@ -3,7 +3,7 @@ package vsys.account
 import org.scalatest.prop.{GeneratorDrivenPropertyChecks, PropertyChecks}
 import org.scalatest.{Matchers, PropSpec}
 
-class AccountOrAliasTests extends PropSpec with PropertyChecks with GeneratorDrivenPropertyChecks with Matchers {
+class AddressOrAliasTests extends PropSpec with PropertyChecks with GeneratorDrivenPropertyChecks with Matchers {
 
   property("Account should get parsed correctly") {
     AddressOrAlias.fromString("ATxt5tn81hXmEaCzuGB7WbJjNfrtRfsAUyt").right.get shouldBe an[Address]
